[mini] inline build to save a few more bytes
